Make the new python backend the default.
- Moves the previous python2/register.py to
  python/register.py, so the backend names make sense.

- Creates a python_old/ backend with the old registration.

To use the old backend, do something like this in pants.ini:

backend_packages: +[
     "pants.backend.python_old",
  ],-["pants.backend.python"]

If you do so you will see a deprecation warning.
