+++ pants.ini
@@ -37,7 +37,6 @@ pythonpath: [
   ]
 
 backend_packages: +[
-    "pants.backend.python2",
     "pants.backend.docgen",
     "internal_backend.optional",
     "internal_backend.repositories",
@@ -52,7 +51,7 @@ backend_packages: +[
     "pants.contrib.node",
     "pants.contrib.python.checks",
     "pants.contrib.scrooge",
-  ],-["pants.backend.python"]
+  ]
 
 # Path patterns to ignore for filesystem operations on top of the builtin patterns.
 pants_ignore: +[
