+++ src/python/pants/backend/python/register.py
@@ -13,21 +13,15 @@ from pants.backend.python.targets.python_binary import PythonBinary
 from pants.backend.python.targets.python_library import PythonLibrary
 from pants.backend.python.targets.python_requirement_library import PythonRequirementLibrary
 from pants.backend.python.targets.python_tests import PythonTests
+from pants.backend.python.tasks.python_isort import IsortPythonTask
 from pants.backend.python.tasks2.gather_sources import GatherSources
-from pants.backend.python.tasks2.pytest_run import PytestRun as PytestRun2
-from pants.backend.python.tasks2.python_binary_create import \
-  PythonBinaryCreate as PythonBinaryCreate2
-from pants.backend.python.tasks2.python_repl import PythonRepl as PythonRepl2
-from pants.backend.python.tasks2.python_run import PythonRun as PythonRun2
+from pants.backend.python.tasks2.pytest_run import PytestRun
+from pants.backend.python.tasks2.python_binary_create import PythonBinaryCreate
+from pants.backend.python.tasks2.python_repl import PythonRepl
+from pants.backend.python.tasks2.python_run import PythonRun
 from pants.backend.python.tasks2.resolve_requirements import ResolveRequirements
 from pants.backend.python.tasks2.select_interpreter import SelectInterpreter
-from pants.backend.python.tasks2.setup_py import SetupPy as SetupPy2
-from pants.backend.python.tasks.pytest_run import PytestRun
-from pants.backend.python.tasks.python_binary_create import PythonBinaryCreate
-from pants.backend.python.tasks.python_isort import IsortPythonTask
-from pants.backend.python.tasks.python_repl import PythonRepl
-from pants.backend.python.tasks.python_run import PythonRun
-from pants.backend.python.tasks.setup_py import SetupPy
+from pants.backend.python.tasks2.setup_py import SetupPy
 from pants.build_graph.build_file_aliases import BuildFileAliases
 from pants.build_graph.resources import Resources
 from pants.goal.task_registrar import TaskRegistrar as task
@@ -55,18 +49,19 @@ def build_file_aliases():
 
 
 def register_goals():
-  task(name='python-binary-create', action=PythonBinaryCreate).install('binary')
-  task(name='pytest', action=PytestRun).install('test')
-  task(name='py', action=PythonRun).install('run')
-  task(name='py', action=PythonRepl).install('repl')
-  task(name='setup-py', action=SetupPy).install()
-  task(name='isort', action=IsortPythonTask).install('fmt')
-
   task(name='interpreter', action=SelectInterpreter).install('pyprep')
   task(name='requirements', action=ResolveRequirements).install('pyprep')
   task(name='sources', action=GatherSources).install('pyprep')
-  task(name='py', action=PythonRun2).install('run2')
-  task(name='pytest', action=PytestRun2).install('test2')
-  task(name='py', action=PythonRepl2).install('repl2')
-  task(name='py', action=PythonBinaryCreate2).install('binary2')
-  task(name='setup-py2', action=SetupPy2).install()
+  task(name='py', action=PythonRun).install('run')
+  task(name='pytest', action=PytestRun).install('test')
+  task(name='py', action=PythonRepl).install('repl')
+  task(name='setup-py', action=SetupPy).install()
+  task(name='py', action=PythonBinaryCreate).install('binary')
+  task(name='isort', action=IsortPythonTask).install('fmt')
+
+  # Integration tests of the new pipeline still use the `*2` aliases.
+  task(name='py', action=PythonRun).install('run2')
+  task(name='pytest', action=PytestRun).install('test2')
+  task(name='py', action=PythonRepl).install('repl2')
+  task(name='py', action=PythonBinaryCreate).install('binary2')
+  task(name='setup-py2', action=SetupPy).install()
